fixes #254, don't use 64-bit sized ints so wasm-bindgen doesn't generate BigInt64Array (#267)
references for better browser support

+++ crates/entities/src/lib.rs
@@ -8,7 +8,7 @@ use sea_orm::{DatabaseConnection, DbBackend, DbErr, FromQueryResult, Statement};
 use shared::response::LibraryStats;
 #[derive(Debug, FromQueryResult)]
 pub struct CountByStatus {
-    count: i64,
+    count: i32,
     name: String,
     status: String,
 }
